requirements.txt
aiohttp
APScheduler
astroid
async-timeout
attrs
beautifulsoup4
chardet
click
discord.py
idna
isort
joblib
lazy-object-proxy
mccabe
multidict
nltk
pytz
regex
six
soupsieve
textblob
toml
tqdm
tzlocal
websockets
wrapt
yarl
